Spark/requirements.txt
# Road Risk Pipeline Dependencies

# Core dependencies
pyspark>=3.3.0
pyarrow>=10.0.0
pandas>=1.5.0

# Configuration
pyyaml>=6.0

# Airflow (optional - for orchestration)
apache-airflow>=2.5.0
apache-airflow-providers-apache-spark>=4.0.0

# Testing (optional)
pytest>=7.0.0
pytest-spark>=0.6.0

# Code quality (optional)
black>=22.0.0
pylint>=2.15.0
